Langchain/chatbot/requirements.txt
langchain==0.3.15
langchain-community==0.3.15
python-dotenv==1.0.0
streamlit==1.47.0